CLOUDSTACK-3797: Builtin appliance definition using veewee
This appliance will be used for testing purposes only. The appliance
will replace the built-in in a test environment before deployment.

Some of the integration tests require specific services running within
guests to test for cloudstack features. Eg: VPC LB needs an httpd server
running. When we need to customize the template this appliance can be
used instead of having custom templates for each test.

Signed-off-by: Prasanna Santhanam <tsp@apache.org>