Revert "fix CLOUDSTACK-2989, do not list network offering with specifyvlan=true for non-root-admin user"
This reverts commit 94254e3c2bd38573adac22d0c66817db3cfa1873.

Conflicts:
	server/src/com/cloud/configuration/ConfigurationManagerImpl.java

Regular user should be able to list the network offering with specifyVlan=true. He just can't use when create the network as he is unaware of the physical network topology. But as the admin can create the network on the user's behalf using this kind of offering, so the user should be able to see the offering's details