Removing coded documentation
This is the output of the program, the developer documentation may be a better place

-----

When developing against the CloudStack Orchestration Platform, you must following the following rules:
API Rule: Always be prepared to handle RuntimeExceptions.

When writing APIs, you must follow these rules:
API Rule: You may think you're the greatest developer in the world but every change to the API must be reviewed and approved.
API Rule: Every API must have unit tests written against it.  And not it's unit tests
API Rule:

-----

Signed-off-by: Laszlo Hornyak <laszlo.hornyak@gmail.com>