Adding new IAM service under services. There are two modules to this component:
part a) plugin - This contains all adapter implementations and components necessary to interact with CloudStack.

part b) server - This contains IAM interfaces that are not aware of any CloudStack entities, they are pure IAM interface that can be implemented in varied ways.
The default implementation provided just works on DB and generic Dao layer.
But these interfaces can be implemented using connectors interacting with LDAP or AD.