Automation of CCP Objects Verification after external changes made to the original setup Purpose of this code:
Generate CCP Objects (VMs, Volumes, Snapshots, VPC, etc..) and CCP Use Cases (Networking, Data Content,etc) before an external action on the CCP Setup and verify the integrity of the CCP Objects and the Use Cases after the external action on the CCP Setup. The integrity of the CCP Objects is verified by performing operations that test the Usability of the objects. This validates the intactness of the setup after an external action. The submitted patch covers only few major use cases. It proves that similar code can be added in future to address similar goals in verifying the integrity of CCP objects belonging to different components of the product.

The code format can be followed to verify validity of real time business use cases while any code changes (CCP,hypervisor,external devices code, etc…) happen over a period of time.

The following are the scenarios that the code format can be used for:
1.Upgrade Validity Verification
     a. CCP Upgrade
     b. Hypervisors Upgrade
     c. External Devices Upgrade
     d. System VM Template Changes.
2.Patch Validity Verification

Code can be used as one of the primary Components to validity Upgrades. It will facilitate the automation of Upgrade Test Verification completely.

How to use the code:

*Kindly make the corresponding substitutions in the commands listed below.

Execute:
 nosetests --with-marvin --marvin-config=$CONFIG $BASEDIR/integration/component/ test_minimal_ug_check.py --load -a tags=preupgrade

After Upgrade or any Changes done to the Setup, Verify that the existing CCP objects are not affected due to the external changes.

Execute:
nosetests --with-marvin --marvin-config=$CONFIG $BASEDIR/integration/component/ test_minimal_ug_check.py --load -a tags=postupgrade

Signed-off-by: SrikanteswaraRao Talluri <talluri@apache.org>